simple zen opener finished
