--- 31_dhtml_parser/cv_updater_cli.py
@@ -1,8 +1,0 @@
-import requests
-import bs4
-
-def main():
-	return None
-
-if __name__ == '__main__':
-	main()
