--- 31_dhtml_parser/updater/bin/pip3.7
@@ -1,10 +1,0 @@
-#!/Users/pepa-pcx/Documents/development/routine_python/15_selenium_web_driver/updater/bin/python3
-# -*- coding: utf-8 -*-
-import re
-import sys
-
-from pip._internal import main
-
-if __name__ == '__main__':
-    sys.argv[0] = re.sub(r'(-script\.pyw?|\.exe)?$', '', sys.argv[0])
-    sys.exit(main())
