--- 31_dhtml_parser/updater/bin/python
@@ -1,1 +1,0 @@
-python3
