--- 31_dhtml_parser/updater/bin/python3
@@ -1,1 +1,0 @@
-/Library/Developer/CommandLineTools/usr/bin/python3
